fix(clp-package): Align single-container runtime config
