+++ tools/deployment/package/single-container/supervisord.conf
@@ -65,7 +65,7 @@ stderr_logfile=/var/log/clp-single-container/results-cache-indices-creator.stder
 user=clp-user
 
 [program:compression-scheduler]
-command=/usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done -- /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u -m job_orchestration.scheduler.compress.compression_scheduler --config %(ENV_CLP_CONFIG_PATH)s
+command=/bin/bash -c 'export RESULT_BACKEND="redis://default:${CLP_REDIS_PASS:?Please set CLP_REDIS_PASS.}@redis:${CLP_REDIS_CONNECT_PORT}/${CLP_REDIS_BACKEND_DB_COMPRESSION:-1}"; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done -- /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u -m job_orchestration.scheduler.compress.compression_scheduler --config %(ENV_CLP_CONFIG_PATH)s'
 autorestart=true
 priority=110
 stdout_logfile=/var/log/clp-single-container/compression-scheduler.stdout.log
@@ -73,7 +73,7 @@ stderr_logfile=/var/log/clp-single-container/compression-scheduler.stderr.log
 user=clp-user
 
 [program:compression-worker]
-command=/usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u /opt/clp/lib/python3/site-packages/bin/celery -A job_orchestration.executor.compress worker --concurrency %(ENV_CLP_COMPRESSION_WORKER_CONCURRENCY)s --loglevel WARNING -f /var/log/compression_worker/worker.log -Q compression -n compression-worker@%%h
+command=/bin/bash -c 'export RESULT_BACKEND="redis://default:${CLP_REDIS_PASS:?Please set CLP_REDIS_PASS.}@redis:${CLP_REDIS_CONNECT_PORT}/${CLP_REDIS_BACKEND_DB_COMPRESSION:-1}"; exec /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u /opt/clp/lib/python3/site-packages/bin/celery -A job_orchestration.executor.compress worker --concurrency %(ENV_CLP_COMPRESSION_WORKER_CONCURRENCY)s --loglevel WARNING -f /var/log/compression_worker/worker.log -Q compression -n compression-worker@%%h'
 autorestart=true
 priority=120
 stdout_logfile=/var/log/clp-single-container/compression-worker.stdout.log
@@ -81,7 +81,7 @@ stderr_logfile=/var/log/clp-single-container/compression-worker.stderr.log
 user=clp-user
 
 [program:query-scheduler]
-command=/usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done -- /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u -m job_orchestration.scheduler.query.query_scheduler --config %(ENV_CLP_CONFIG_PATH)s
+command=/bin/bash -c 'export RESULT_BACKEND="redis://default:${CLP_REDIS_PASS:?Please set CLP_REDIS_PASS.}@redis:${CLP_REDIS_CONNECT_PORT}/${CLP_REDIS_BACKEND_DB_QUERY:-0}"; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done -- /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u -m job_orchestration.scheduler.query.query_scheduler --config %(ENV_CLP_CONFIG_PATH)s'
 autorestart=true
 priority=130
 stdout_logfile=/var/log/clp-single-container/query-scheduler.stdout.log
@@ -89,7 +89,7 @@ stderr_logfile=/var/log/clp-single-container/query-scheduler.stderr.log
 user=clp-user
 
 [program:query-worker]
-command=/usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u /opt/clp/lib/python3/site-packages/bin/celery -A job_orchestration.executor.query worker --concurrency %(ENV_CLP_QUERY_WORKER_CONCURRENCY)s --loglevel WARNING -f /var/log/query_worker/worker.log -Q query -n query-worker@%%h
+command=/bin/bash -c 'export RESULT_BACKEND="redis://default:${CLP_REDIS_PASS:?Please set CLP_REDIS_PASS.}@redis:${CLP_REDIS_CONNECT_PORT}/${CLP_REDIS_BACKEND_DB_QUERY:-0}"; exec /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 5672 /usr/local/bin/clp-single-container-wait-for-tcp 127.0.0.1 6379 python3 -u /opt/clp/lib/python3/site-packages/bin/celery -A job_orchestration.executor.query worker --concurrency %(ENV_CLP_QUERY_WORKER_CONCURRENCY)s --loglevel WARNING -f /var/log/query_worker/worker.log -Q query -n query-worker@%%h'
 autorestart=true
 priority=140
 stdout_logfile=/var/log/clp-single-container/query-worker.stdout.log
@@ -104,9 +104,20 @@ stdout_logfile=/var/log/clp-single-container/reducer.stdout.log
 stderr_logfile=/var/log/clp-single-container/reducer.stderr.log
 user=clp-user
 
+[program:garbage-collector]
+command=/bin/bash -c 'if [[ "${CLP_GARBAGE_COLLECTOR_ENABLED:-1}" == "0" ]]; then echo "garbage-collector disabled."; exit 0; fi; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done /run/clp-single-container/results-cache-indices-creator.done -- python3 -u -m job_orchestration.garbage_collector.garbage_collector --config %(ENV_CLP_CONFIG_PATH)s'
+autorestart=unexpected
+startsecs=0
+priority=155
+stdout_logfile=/var/log/clp-single-container/garbage-collector.stdout.log
+stderr_logfile=/var/log/clp-single-container/garbage-collector.stderr.log
+user=clp-user
+environment=CLP_LOGGING_LEVEL="%(ENV_CLP_GARBAGE_COLLECTOR_LOGGING_LEVEL)s",CLP_LOGS_DIR="/var/log/garbage_collector"
+
 [program:api-server]
-command=/usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done /run/clp-single-container/results-cache-indices-creator.done -- /opt/clp/bin/api_server --host 0.0.0.0 --port 3001 --config %(ENV_CLP_CONFIG_PATH)s
-autorestart=true
+command=/bin/bash -c 'if [[ "${CLP_API_SERVER_ENABLED:-1}" == "0" ]]; then echo "api-server disabled."; exit 0; fi; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done /run/clp-single-container/results-cache-indices-creator.done -- /opt/clp/bin/api_server --host 0.0.0.0 --port 3001 --config %(ENV_CLP_CONFIG_PATH)s'
+autorestart=unexpected
+startsecs=0
 priority=160
 stdout_logfile=/var/log/clp-single-container/api-server.stdout.log
 stderr_logfile=/var/log/clp-single-container/api-server.stderr.log
@@ -114,7 +125,7 @@ user=clp-user
 
 [program:log-ingestor]
 command=/bin/bash -c 'if [[ "${CLP_LOG_INGESTOR_ENABLED:-0}" == "0" ]]; then echo "log-ingestor disabled."; exit 0; fi; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done -- /opt/clp/bin/log-ingestor --config %(ENV_CLP_CONFIG_PATH)s --host 0.0.0.0 --port 3002'
-autorestart=false
+autorestart=unexpected
 startsecs=0
 priority=165
 stdout_logfile=/var/log/clp-single-container/log-ingestor.stdout.log
@@ -124,7 +135,7 @@ environment=RUST_LOG="%(ENV_CLP_LOG_INGESTOR_LOGGING_LEVEL)s"
 
 [program:mcp-server]
 command=/bin/bash -c 'if [[ "${CLP_MCP_SERVER_ENABLED:-0}" == "0" ]]; then echo "MCP server disabled."; exit 0; fi; exec /usr/local/bin/clp-single-container-wait-for-files /run/clp-single-container/db-table-creator.done /run/clp-single-container/results-cache-indices-creator.done -- python3 -u -m clp_mcp_server.clp_mcp_server --host 0.0.0.0 --port 8000 --config-path %(ENV_CLP_CONFIG_PATH)s'
-autorestart=false
+autorestart=unexpected
 startsecs=0
 priority=168
 stdout_logfile=/var/log/clp-single-container/mcp-server.stdout.log
